implement single-pass install with script
