+++ brewblox_ctl/commands/snapshot.py
@@ -2,7 +2,7 @@
 Brewblox-ctl snapshot commands
 """
 
-from os import listdir, path
+from pathlib import Path
 from tempfile import TemporaryDirectory
 
 import click
@@ -21,14 +21,13 @@ def snapshot():
 
 
 @snapshot.command()
-@click.option('--dir',
-              help='Brewblox directory')
 @click.option('--file',
-              help='Snapshot file')
+              help='Snapshot file',
+              default='../brewblox.tar.gz')
 @click.option('--force',
               is_flag=True,
               help='Remove previous tarfile if it exists')
-def save(dir, file, force):
+def save(file, force):
     """Save Brewblox directory to snapshot.
 
     This can be used to move Brewblox installations between hosts.
@@ -37,19 +36,9 @@ def save(dir, file, force):
 
     Block data stored on Spark controllers is not included in the snapshot.
     """
+    utils.check_config()
     utils.confirm_mode()
-
-    if utils.is_brewblox_cwd():
-        dir = dir or '.'
-        file = file or '../brewblox.tar.gz'
-    else:
-        dir = dir or './brewblox'
-        file = file or './brewblox.tar.gz'
-
-    dir = path.abspath(dir)
-
-    if not utils.is_brewblox_dir(dir):
-        raise ValueError(f'`{dir}` is not a Brewblox directory')
+    dir = Path('./').resolve()
 
     if utils.path_exists(file):
         if force or utils.confirm(f'`{file}` already exists. ' +
@@ -58,51 +47,29 @@ def save(dir, file, force):
         else:
             return
 
-    basedir = path.basename(dir)
-    parent = dir + '/..'
-    sh(f'sudo tar -C {parent} -czf {file} {basedir}')
-    click.echo(path.abspath(path.expanduser(file)))
+    basedir = dir.name
+    sh(f'sudo tar -C {dir.parent} -czf {file} {basedir}')
+    click.echo(Path.resolve(file))
 
 
 @snapshot.command()
-@click.option('--dir',
-              help='Brewblox directory')
 @click.option('--file',
-              help='Snapshot file')
-@click.option('--force',
-              is_flag=True,
-              help='Remove directory if it exists')
-def load(dir, file, force):
+              help='Snapshot file',
+              default='../brewblox.tar.gz')
+def load(file):
     """Create Brewblox directory from snapshot.
 
     This can be used to move Brewblox installations between hosts.
     To create a snapshot, use `brewblox-ctl snapshot save`
     """
+    utils.check_config()
     utils.confirm_mode()
-
-    if utils.is_brewblox_cwd():
-        dir = dir or '.'
-        file = file or '../brewblox.tar.gz'
-    else:
-        dir = dir or './brewblox'
-        file = file or './brewblox.tar.gz'
-
-    dir = path.abspath(dir)
-
-    if utils.path_exists(dir) and not utils.is_empty_dir(dir):
-        if not utils.is_brewblox_dir(dir):
-            raise FileExistsError(f'`{dir}` is not a Brewblox directory.')
-        if force or utils.confirm(f'`{dir}` already exists. ' +
-                                  'Do you want to continue and erase its content?'):
-            # we'll do the actual rm after unpacking files
-            utils.info(f'Contents of `{dir}` will be removed')
-        else:
-            return
+    dir = Path('./').resolve()
 
     with TemporaryDirectory() as tmpdir:
         utils.info(f'Extracting snapshot to {dir} directory...')
         sh(f'tar -xzf {file} -C {tmpdir}')
-        content = listdir(tmpdir)
+        content = list(Path(tmpdir).iterdir())
         if utils.ctx_opts().dry_run:
             content = ['brewblox']
         if len(content) != 1:
