+++ brewblox_ctl/const.py
@@ -1,6 +1,7 @@
 """
 Const values
 """
+import pathlib
 import sys
 
 PY = sys.executable
@@ -8,8 +9,11 @@ ARGS = sys.argv
 CLI = f'{PY} -m brewblox_ctl'
 HOST = 'https://localhost'
 
+DOCKER = 'docker'
+COMPOSE = '.venv/bin/docker-compose'
+
 CFG_VERSION_KEY = 'BREWBLOX_CFG_VERSION'
-LIB_RELEASE_KEY = 'BREWBLOX_CTL_LIB_RELEASE'
+CTL_RELEASE_KEY = 'BREWBLOX_CTL_RELEASE'
 RELEASE_KEY = 'BREWBLOX_RELEASE'
 SKIP_CONFIRM_KEY = 'BREWBLOX_SKIP_CONFIRM'
 DEBUG_KEY = 'BREWBLOX_DEBUG'
@@ -27,8 +31,10 @@ LOG_INFO = 'INFO'.ljust(10)
 LOG_WARN = 'WARN'.ljust(10)
 LOG_ERR = 'ERROR'.ljust(10)
 
-DATA_DIR = './brewblox_ctl_lib/data'
-CONFIG_DIR = f'{DATA_DIR}/config'
+CTL_DIR = str(pathlib.Path(__file__).parent.resolve())
+DEPLOY_DIR = f'{CTL_DIR}/deployed'
+CONFIG_DIR = f'{DEPLOY_DIR}/config'
+SCRIPT_DIR = f'{DEPLOY_DIR}/scripts'
 AVAHI_CONF = '/etc/avahi/avahi-daemon.conf'
 UI_DATABASE = 'brewblox-ui-store'
