+++ brewblox_ctl/deployed/config/docker-compose.shared.yml
@@ -67,3 +67,8 @@ services:
     restart: unless-stopped
     labels:
       - traefik.http.routers.ui.rule=PathPrefix(`/ui`) || Path(`/`)
+
+networks:
+  default:
+    driver_opts: 
+      com.docker.network.bridge.name: br-${COMPOSE_PROJECT_NAME}
