+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "brewblox-ctl"
-version = "0.27.0"
+version = "1.0.0"
 description = "Brewblox management tool"
 authors = ["BrewPi <development@brewpi.com>"]
 license = "GPL-3.0"
@@ -11,13 +11,14 @@ python = ">=3.6,<4"
 requests = "^2.26.0"
 click = ">=7"
 python-dotenv = {extras = ["cli"], version = "^0.19.0"}
-docker = "^5.0.0"
-docker-compose = "^1.29.2"
 configobj = "^5.0.6"
 zeroconf = "^0.33.1"
 PyYAML = "^5.4.1"
 pyusb = ">=1.2"
 pyserial = ">=3"
+docker = "^5.0.3"
+docker-compose = "^1.29.2"
+websocket-client = ">=0.32.0,<1"
 
 [tool.poetry.dev-dependencies]
 pytest-cov = "^2.12.1"
@@ -29,9 +30,6 @@ pathlib2 = "^2.3.6"
 flake8-quotes = "^3.2.0"
 httpretty = "1.0.5"
 importlib-metadata = "^4.8.1"
-
-[tool.poetry.scripts]
-brewblox-ctl = 'brewblox_ctl.__main__:main'
 
 [build-system]
 requires = ["poetry>=0.12"]
