+++ tox.ini
@@ -1,6 +1,7 @@
 [pytest]
 flake8-max-line-length = 120
 addopts =
+    --ignore=brewblox/
     --cov=brewblox_ctl
     --cov-branch
     --cov-report=term-missing:skip-covered
